Added a second display line
After having created a Project branch called DisplayDate off of Master, I wanted to start making changes in this branch.  This is the first edit so that DisplayDate will now be divergent from Master and then I will eventually RI my code into Master for a compile.  Eventually this will be set up for a CI process.